// src/common/types/patient.type.ts
export interface Patient {
  id: string;
  first_name: string;
  last_name: string;
  sex: string;
  date_of_birth: string;
  email_address: string;
  health_card_number: string;
  phone_number: string;
  age: string;
  preferred_clinic: any;
  most_visited_clinic: any;
  next_appointment_date: string;
  
}